% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bray.curtis.R
\name{bray.curtis}
\alias{bray.curtis}
\title{Bray-Curtis dissimilarity}
\usage{
bray.curtis(input.data, target.plot.ID)
}
\arguments{
\item{input.data}{A species x relevés matrix with abundance values}

\item{target.plot.ID}{ID or label of target plot}
}
\value{
A vector of Bray-Curtis dissimilarities between the target and all the other relevés in input.data
}
\description{
Calculates Bray-Curtis dissimilarity between a target plot, and all the other relevés in the dataset
}
\author{
Francesco Maria Sabatini

Helge Bruelheide
}
