% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mij.calc.R
\name{Mij.calc}
\alias{Mij.calc}
\title{Pairwise likelihood of species co-occurrence}
\usage{
Mij.calc(dt, verbose = F)
}
\arguments{
\item{dt}{A data.frame with three columns: releve ID, species ID, and abundance}
}
\value{
Returns a square matrix of the likelihood of pairwise species co-occurrence
}
\description{
Intermediate step to calculate Beals' smoothing
}
\author{
Francesco Maria Sabatini

Helge Bruelheide
}
