% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAR.IIIb.R
\name{SAR.IIIb}
\alias{SAR.IIIb}
\title{Rarefaction curves}
\usage{
SAR.IIIb(x, areas, n = 10)
}
\arguments{
\item{x}{A vector of species richness}

\item{areas}{A vector of plot sizes}
}
\value{
Returns a dataframe of cumulative richness and pooled plot areas
}
\description{
Builds an empirical rarefaction curve
}
\author{
Francesco Maria Sabatini

Helge Bruelheide
}
