% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeciesPool_sPlot.R
\name{SpeciesPool_sPlot}
\alias{SpeciesPool_sPlot}
\title{Species pool based on Beal's smoothing}
\usage{
SpeciesPool_sPlot(input.data, coords, Mij = NULL, ncores = 1,
  rows = NULL, t.radius = 20000, t.bray = 0.2, t.plot.number = 10L,
  verbose = T, species.list = F)
}
\arguments{
\item{input.data}{data.frame of species abundances across relevés. It should have three columns: Relevé ID, Species ID, and species abundance}

\item{coords}{A SpatialPointsDataFrame with the geographic coordinates of all plots. It should have Relevè ID and areas defined in the @data}

\item{Mij}{Matrix of pairwise likelihood of species co-occurrence (Sparse matrices accepted)}

\item{ncores}{integer indicating the number of cores to use. If ncores>1 the calculation will be don in parallel}

\item{rows}{a vector of integers indicating on which rows of input.data the function should run}

\item{t.radius}{Threshold of geographic buffer around target relevé}

\item{t.bray}{Threshold of bray-curtis dissimilarity for selection relevé compositionally similar to target relevé}

\item{t.plot.number}{Minimum number of neighbouring relevés for calculating rarefaction curves}

\item{verbose}{logical}

\item{species.list}{logical - Should the list of species composing the species pool be returned?}
}
\value{
Returns a dataframe containing for each relevé:
Species - the number of species observed across all relevé neighouting the target relevé
Chao, iChao2, jack1, jack2  - various species richness estimates, standard errors, as derived from the function SpadeR::ChaoSpecies
nplots - number of relevés within a t.radius distance from the target relevé having a bray-curtis dissimilarity lower than t.bray
beals.at.chao - cut-off of Beals' occurrence likelihood, selected as the ith species corresponding to chao
n.plots.area - number of relevés within a t.radius distance, and t.bray dissimilarity from target relevés for which area data is available
arr, gomp, mm, Asymp - parameter estimates for different empirical non-linear functions fitted to rarefaction curves, with relative AIC
sp.pool.list - list of species compatible with target relevé, i.e. having a Beals' likelihood lower than beals.at.chao
}
\description{
For each relevé, this function selectes all neighbouring relevé having a similar potential species composition, and fits empirical non-linear functions to rarefaction curves.
}
\author{
Francesco Maria Sabatini

Helge Bruelheide
}
