% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beals.all.R
\name{beals.all}
\alias{beals.all}
\title{Beals' smoothing}
\usage{
beals.all(x.names, x.cover, Mij)
}
\arguments{
\item{x.names}{vector of species names or labels}

\item{x.cover}{vector of species presence-absences}

\item{Mij}{matrix of pairwise likelihood of species co-occurrence (sparse matrices accepted)}
}
\value{
A vector of Bray-Curtis dissimilarity between the target plot and all the other relevés in input.data
}
\description{
Calculates Beals' smoothing for all species in a target plot, based on Mij matrix
}
\references{
Ewald, J. (2002) Multiple controls of understorey plant richness in mountain forests of the Bavarian Alps. Phytocoenologia, 32, 85-100.
}
\author{
Francesco Maria Sabatini

Helge Bruelheide
}
