% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeciesPool.R
\name{SpeciesPool}
\alias{SpeciesPool}
\title{Species pool based on Beal's smoothing}
\usage{
SpeciesPool(input.data, coords, Mij = NULL, ncores = 1, rows = NULL,
  t.radius = 20000, t.bray = 0.2, t.plot.number = 10L,
  cutoff = c("iChao2", "Gompertz", "Michaelis"), verbose = T,
  species.list = F)
}
\arguments{
\item{input.data}{data.frame of species abundances across relevés. It should have three columns: one with Relevé IDs, one with Species ID, and one with species abundance/cover values}

\item{coords}{A SpatialPointsDataFrame with the geographic coordinates of all plots. It should have Relevé IDs and areas defined in the data}

\item{Mij}{matrix of pairwise likelihood of species co-occurrence (sparse matrices accepted). If not provided, it will be calculated from the data}

\item{ncores}{integer indicating the number of cores to use. If ncores>1 the calculation will be done in parallel}

\item{rows}{a vector of integers indicating on which rows of the input.data the function should run}

\item{t.radius}{threshold of geographic buffer around target relevé}

\item{t.bray}{threshold of Bray-Curtis dissimilarity for selecting relevés compositionally similar to target relevé}

\item{t.plot.number}{minimum number of neighbouring relevés for calculating rarefaction curves}

\item{cutoff}{method used to estimate the size of the species pool. Default is 'iChao2', other possible are 'Gompertz' or 'Michaelis'}

\item{verbose}{logical}

\item{species.list}{logical: Should the list of species composing the species pool be returned?}
}
\value{
Returns a dataframe containing for each relevé:
- Species -  the number of species observed across all relevés neighouring the target relevé\cr
- Chao, iChao2, jack1, jack2 -  various species richness estimates and standard errors, as derived from the function SpadeR::ChaoSpecies\cr
- nplots -  number of relevés within a t.radius distance from the target relevé having a bray-curtis dissimilarity lower than t.bray\cr
- beals.at.chao -  cut-off of Beals' occurrence likelihood, selected as the ith species corresponding to chao\cr
- n.plots.area -   number of relevés within a t.radius distance, and t.bray dissimilarity from target relevés for which area data is available\cr
- arr, gomp, mm, Asymp -   parameter estimates for different empirical non-linear functions fitted to rarefaction curves, with relative AIC\cr
- sp.pool.list -   list of species compatible with target relevé, i.e. having a Beals' likelihood lower than beals.at.chao\cr
}
\description{
For each relevé, this function selects all neighbouring relevés having a similar potential species composition, and fits empirical non-linear functions to rarefaction curves.
}
\examples{
load(GVRD_test_data.RData) ## data provided separately, make sure it's in working directory
aa <- SpeciesPool(DT.fb, mycoords, ncores=3, rows=1:100, t.radius=20000,
>         t.bray=0.2, t.plot.number=10L, verbose=T, species.list = T)
}
\author{
Francesco Maria Sabatini

Helge Bruelheide
}
